230614/정윤석/패시브 특성 구현
## 구현 목록

1. TraitManager에서 선택한 특성들 중 패시브 특성을 캐릭터 데이터에 적용하도록 설정함.(레벨은 20레벨을 기준으로 두었으며 특성은 각 레벨당 선택할 수 있는 모든 특성에서 1번째를 선택했다고 가정한다.)

## 자잘한 수정 사항들

1. traitTrash에서 traitEffectType을 Enum으로 설정하고 이를 Parse할 수 있도록 설정(대충 코드를 좀 보기 쉽게 수정했다는 뜻 + 아래 표 참조)

+++ Assets/Scenes/Night/Script/TrashData/CharacterTrashData.cs
@@ -34,7 +34,7 @@ public class CharacterTrashData
 
     public int currentMoney;                //보유 재화
     public int startMoney;                  //시작 재화
-    public int earnMoney;                   //획득 재화
+    public float earnMoney;                   //획득 재화
 
     public int shopSlot;                    //상점 슬롯
     public int shopMinRank;                 //상점 최소 랭크
